requirements.txt
# Core FastAPI and server dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3
pydantic==2.10.4
pydantic-core==2.27.2
pydantic-settings==2.7.1

# Async support
anyio==4.7.0
asyncio-mqtt
uvloop==0.21.0

# Database dependencies
sqlalchemy==2.0.36
sqlalchemy-utils==0.41.2
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
aiosqlite

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0
PyJWT==2.10.1
python-multipart>=0.0.6
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0

# Third-party authentication
fastapi-sso==0.17.0
oauthlib==3.2.2

# Session and middleware
itsdangerous==2.2.0

# Environment and configuration
python-dotenv==1.0.1
pydantic-settings==2.7.1

# File handling and storage
aiofiles==24.1.0
minio==7.2.14
boto3>=1.15.0
Pillow>=9.0.0
python-magic==0.4.27

# HTTP clients and networking
httpx==0.28.1
httpcore==1.0.7
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiohappyeyeballs==2.4.6
requests==2.32.3
urllib3==2.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10

# Email services
sendgrid==6.11.0
python-http-client==3.3.7

# SMS services
twilio==8.10.0

# Database utilities
motor==3.6.0
pymongo==4.9.2
beanie==1.28.0
dnspython==2.7.0

# Caching (if using Redis)
redis

# Utilities
click==8.1.8
loguru==0.7.3
python-dateutil==2.9.0.post0
six==1.17.0
toml==0.10.2
PyYAML==6.0.2
fuzzywuzzy==0.18.0

# Cryptography and encoding
ecdsa==0.19.0
starkbank-ecdsa==2.2.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
rsa==4.9

# Development and testing
pytest
pytest-asyncio
httpx  # for testing FastAPI

# System utilities
psutil==6.1.1
pathlib2==2.3.7.post1

# Permissions and authorization
permit==2.7.2

# Video processing
moviepy==2.1.2
opencv-python==4.11.0.86

# Data processing
numpy==2.2.2

# Web utilities
websockets==14.1
watchfiles==1.0.3

# Additional utilities
attrs==25.1.0
frozenlist==1.5.0
greenlet==3.2.1
h11==0.14.0
httptools==0.6.4
lazy-model==0.2.0
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
propcache==0.3.0
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.18.3
zipp==3.21.0

# Jupyter support (if needed for development)
ipykernel==6.29.5
ipython==8.31.0
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7

# Additional development tools
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
jedi==0.19.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyzmq==26.2.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
wcwidth==0.2.13

# Email validation
email_validator==2.2.0

# Additional utilities
backcall==0.2.0
cffi==1.17.1
comm==0.2.2
b2sdk>=1.20.0